Separated test_utilities into separate files. Added nanometers to LengthUnit.
